Share materialized-view transactional command helpers
